+++ src/constants/tagPaths.js
@@ -1,0 +1,4 @@
+export const tagPaths = {
+  allTags: '/tags?limit=100000',
+  alreadyAddedToRecords: '/eholdings/tags/summary?filter[rectype]=provider&filter[rectype]=package&filter[rectype]=title&filter[rectype]=resource',
+};